requirements.txt
Django>=5.1
djangorestframework>=3.15
psycopg2-binary>=2.9
python-decouple
gunicorn
whitenoise>=6.6
django-filter>=24.3   